+++ SolrNet/Impl/ResponseParsers/ResultsResponseParser.cs
@@ -66,8 +66,8 @@ namespace SolrNet.Impl.ResponseParsers {
             if (resultNode == null)
                 return;
 
-            results.NumFound = Convert.ToInt32(resultNode.Attribute("numFound").Value);
-            results.Start = Convert.ToInt32(resultNode.Attribute("start").Value);
+            results.NumFound = Convert.ToInt64(resultNode.Attribute("numFound").Value);
+            results.Start = Convert.ToInt64(resultNode.Attribute("start").Value);
             var maxScore = resultNode.Attribute("maxScore");
             if (maxScore != null)
                 results.MaxScore = double.Parse(maxScore.Value, CultureInfo.InvariantCulture.NumberFormat);
